Fix/circular dependency (#38)
* fix: utils api

* fix: utils api
