+++ libs/api/src/common/index.ts
@@ -1,1 +1,2 @@
 export * from './filters';
+export * from './utils';
